paddleocr-service/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
paddlepaddle==3.0.0
paddleocr>=2.7.0
Pillow>=10.0.0
numpy>=1.24.0,<2.0.0
python-multipart>=0.0.6
